feat(kraken): add connector implementation following Binance pattern
- Add Kraken connector structure (config.py, urm.py, provider stubs)
- Implement OHLCV REST endpoint following Binance pattern
- Remove legacy provider implementation files
- Update providers/kraken/ to use shim pattern like Binance
- URM imports from connectors/kraken/urm.py for consistency
