+++ .gitignore
@@ -21,3 +21,4 @@ dist/
 build/
 
 .coverage
+task.md
