+++ laakhay/data/connectors/kraken/__init__.py
@@ -1,0 +1,3 @@
+"""Kraken connector implementation."""
+
+__all__: list[str] = []
