+++ laakhay/data/connectors/kraken/rest/endpoints/common/__init__.py
@@ -1,0 +1,1 @@
+"""Common Kraken REST endpoints (available for both spot and futures)."""
