+++ laakhay/data/connectors/kraken/rest/endpoints/futures/__init__.py
@@ -1,0 +1,1 @@
+"""Futures-specific Kraken REST endpoints."""
