+++ laakhay/data/connectors/kraken/rest/endpoints/spot/__init__.py
@@ -1,0 +1,1 @@
+"""Spot-specific Kraken REST endpoints."""
