+++ laakhay/data/providers/kraken/rest/__init__.py
@@ -1,5 +1,11 @@
-"""Kraken REST provider."""
+"""Kraken REST provider (shim for backward compatibility).
 
-from .provider import KrakenRESTProvider
+This module will export KrakenRESTProvider from the connector once
+the connector implementation is complete.
+"""
 
-__all__ = ["KrakenRESTProvider"]
+# TODO: Import from connector once connector implementation is complete
+# from laakhay.data.connectors.kraken.rest.provider import KrakenRESTConnector as KrakenRESTProvider
+
+__all__: list[str] = []
+
