+++ laakhay/data/providers/kraken/ws/__init__.py
@@ -1,5 +1,11 @@
-"""Kraken WebSocket provider."""
+"""Kraken WebSocket provider (shim for backward compatibility).
 
-from .provider import KrakenWSProvider
+This module will export KrakenWSProvider from the connector once
+the connector implementation is complete.
+"""
 
-__all__ = ["KrakenWSProvider"]
+# TODO: Import from connector once connector implementation is complete
+# from laakhay.data.connectors.kraken.ws.provider import KrakenWSConnector as KrakenWSProvider
+
+__all__: list[str] = []
+
